+++ src/jsonrpccpp/client/connectors/tcpsocketclient.cpp
@@ -9,7 +9,7 @@
 
 #include "tcpsocketclient.h"
 
-#ifdef __WIN32__
+#ifdef _WIN32
 #include "windowstcpsocketclient.h"
 #else
 #include "linuxtcpsocketclient.h"
@@ -20,7 +20,7 @@ using namespace std;
 
 TcpSocketClient::TcpSocketClient(const std::string &ipToConnect,
                                  const unsigned int &port) {
-#ifdef __WIN32__
+#ifdef _WIN32
   this->realSocket = new WindowsTcpSocketClient(ipToConnect, port);
 #else
   this->realSocket = new LinuxTcpSocketClient(ipToConnect, port);