+++ src/jsonrpccpp/server/connectors/httpserver.cpp
@@ -9,6 +9,7 @@
 
 #include "httpserver.h"
 #include <cstdlib>
+#include <cstring>
 #include <iostream>
 #include <jsonrpccpp/common/specificationparser.h>
 #include <sstream>
@@ -31,6 +32,31 @@ HttpServer::HttpServer(int port, const std::string &sslcert,
     : AbstractServerConnector(), port(port), threads(threads), running(false),
       path_sslcert(sslcert), path_sslkey(sslkey), daemon(NULL) {}
 
+HttpServer::~HttpServer() {
+  if (bind_address != nullptr)
+    freeaddrinfo(bind_address);
+}
+
+bool HttpServer::SetBindAddress(const std::string& addr) {
+  if (bind_address != nullptr) {
+    freeaddrinfo(bind_address);
+    bind_address = nullptr;
+  }
+
+  std::ostringstream port_str;
+  port_str << port;
+
+  const char* addr_cstr = nullptr;
+  if (!addr.empty())
+    addr_cstr = addr.c_str();
+
+  struct addrinfo hints;
+  std::memset(&hints, 0, sizeof(hints));
+  hints.ai_flags = AI_PASSIVE;
+
+  return getaddrinfo(addr_cstr, port_str.str().c_str(), &hints, &bind_address) == 0;
+}
+
 IClientConnectionHandler *HttpServer::GetHandler(const std::string &url) {
   if (AbstractServerConnector::GetHandler() != NULL)
     return AbstractServerConnector::GetHandler();
@@ -48,6 +74,12 @@ bool HttpServer::StartListening() {
     const bool has_poll =
         (MHD_is_feature_supported(MHD_FEATURE_POLL) == MHD_YES);
     unsigned int mhd_flags;
+
+    // If no bind address has been set explicitly, specify one that listens
+    // on all interfaces now.
+    if (bind_address == nullptr && !SetBindAddress(""))
+      return false;
+
     if (has_epoll)
 // In MHD version 0.9.44 the flag is renamed to
 // MHD_USE_EPOLL_INTERNALLY_LINUX_ONLY. In later versions both
@@ -69,6 +101,7 @@ bool HttpServer::StartListening() {
             HttpServer::callback, this, MHD_OPTION_HTTPS_MEM_KEY,
             this->sslkey.c_str(), MHD_OPTION_HTTPS_MEM_CERT,
             this->sslcert.c_str(), MHD_OPTION_THREAD_POOL_SIZE, this->threads,
+            MHD_OPTION_SOCK_ADDR, bind_address->ai_addr,
             MHD_OPTION_END);
       } catch (JsonRpcException &ex) {
         return false;
@@ -76,7 +109,9 @@ bool HttpServer::StartListening() {
     } else {
       this->daemon = MHD_start_daemon(
           mhd_flags, this->port, NULL, NULL, HttpServer::callback, this,
-          MHD_OPTION_THREAD_POOL_SIZE, this->threads, MHD_OPTION_END);
+          MHD_OPTION_THREAD_POOL_SIZE, this->threads,
+          MHD_OPTION_SOCK_ADDR, bind_address->ai_addr,
+          MHD_OPTION_END);
     }
     if (this->daemon != NULL)
       this->running = true;