+++ src/jsonrpccpp/server/connectors/httpserver.h
@@ -15,11 +15,13 @@
 #include <sys/types.h>
 #if defined(_WIN32) && !defined(__CYGWIN__)
 #include <ws2tcpip.h>
+#include <winsock2.h>
 #if defined(_MSC_FULL_VER) && !defined (_SSIZE_T_DEFINED)
 #define _SSIZE_T_DEFINED
 typedef intptr_t ssize_t;
 #endif // !_SSIZE_T_DEFINED */
 #else
+#include <netdb.h>
 #include <unistd.h>
 #include <sys/time.h>
 #include <sys/socket.h>
@@ -49,6 +51,18 @@ namespace jsonrpc
              */
             HttpServer(int port, const std::string& sslcert = "", const std::string& sslkey = "", int threads = 50);
 
+            ~HttpServer();
+
+            /**
+             * Sets the address at which the server will bind when started.
+             * By default, the server listens at all interfaces.  If this
+             * method is called with a non-empty string before starting the
+             * server, then it will only bind to the specified address.
+             * This method returns true on success and false if the given
+             * address could not be set as bind address.
+             */
+            bool SetBindAddress(const std::string& addr);
+
             virtual bool StartListening();
             virtual bool StopListening();
 
@@ -62,6 +76,7 @@ namespace jsonrpc
             int port;
             int threads;
             bool running;
+            struct addrinfo* bind_address = nullptr;
             std::string path_sslcert;
             std::string path_sslkey;
             std::string sslcert;